feat: improvements in structures
